Refactored by using Rx. Code readability improvements.

+++ app/src/main/java/com/example/coinloft/util/ImgUrlFormat.java
@@ -5,6 +5,6 @@ import androidx.annotation.NonNull;
 public interface ImgUrlFormat {
 
     @NonNull
-    String format(int id);
+    String format(long id);
 
 }
